refactored with class based view

+++ posts/api/views.py
@@ -1,23 +1,31 @@
 from django.shortcuts import get_object_or_404
 from rest_framework import status
-from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from posts.models import Post
 from .serializers import PostSerializer
-from .permissions import IsOwnerOrReadOnly
+from . import permissions
 
 
-@api_view(['POST', 'GET'])
-@permission_classes([IsAuthenticated])
-def blog_post(request):
-    if request.method == "GET":
+class BlogPost(APIView):
+    """
+    Create post instance or retrieve all instances.
+    """
+    permission_classes = [IsAuthenticated, permissions.IsOwnerOrReadOnly]
+
+    def get(self, request):
+        """
+        Get all post instances
+        """
         posts = Post.objects.all()
         serializer = PostSerializer(posts, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
-    if request.method == 'POST':
+    def post(self, request):
+        """
+        Retrieve post instance
+        """
         serializer = PostSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save(author=request.user)
@@ -27,23 +35,29 @@ def blog_post(request):
 
 class PostDetail(APIView):
     """
-    Retrieve, update or delete a post instance.
+     Retrieve, update or delete a post instance.
     """
-    permission_classes = [IsAuthenticated, IsOwnerOrReadOnly]
+    permission_classes = [IsAuthenticated, permissions.IsOwnerOrReadOnly]
 
     def get_post(self, slug):
         post = get_object_or_404(Post, slug=slug)
         return post
 
     def get(self, request, slug):
+        """
+        Retrieve post instance by slug
+        """
         post = self.get_post(slug=slug)
         serializer = PostSerializer(post)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def put(self, request, slug):
+        """
+        Update post instance by slug
+        """
         post = self.get_post(slug=slug)
-        # if post.author != request.user:
-        #     return Response({'response': ' only author can edit or delete post'})
+        if post.author != request.user:
+            return Response({'response': ' only author can edit or delete post'})
         serializer = PostSerializer(post, request.data)
         data = {}
         if serializer.is_valid():
@@ -53,33 +67,43 @@ class PostDetail(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, slug):
+        """
+        Delete post instance by slug
+        """
         post = self.get_post(slug=slug)
+        if post.author != request.user:
+            return Response({'response': ' only author can edit or delete post'})
         data = {}
         deleted = post.delete()
         if deleted:
             data['response'] = 'successfully deleted'
         else:
             data['response'] = 'failed to delete'
-        return Response(data=data, status=status.HTTP_204_NO_CONTENT)
+        return Response(data=data)
 
 
-@api_view(['GET', ])
-@permission_classes([IsAuthenticated, ])
-def like(request, slug):
-    post = get_object_or_404(Post, slug=slug)
-    if request.method == 'GET':
-        post.like.add(request.user)
-        post.unlike.remove(request.user)
-        post.save()
-        return Response({'response': 'you liked this post'}, status=status.HTTP_200_OK)
+class Like(APIView):
+    """
+    Like post instance by slug
+    """
+    def get(self, request, slug):
+        post = get_object_or_404(Post, slug=slug)
+        if request.method == 'GET':
+            post.like.add(request.user)
+            post.unlike.remove(request.user)
+            post.save()
+            return Response({'response': 'you like this post'}, status=status.HTTP_200_OK)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
 
 
-@api_view(['GET', ])
-@permission_classes([IsAuthenticated, ])
-def unlike(request, slug):
-    post = get_object_or_404(Post, slug=slug)
-    if request.method == 'GET':
-        post.unlike.add(request.user)
-        post.like.remove(request.user)
-        post.save()
-        return Response({'response': 'you liked this post'}, status=status.HTTP_200_OK)
+class Unlike(APIView):
+    """
+    Unlike post instance by slug
+    """
+    def get(self, request, slug):
+        post = get_object_or_404(Post, slug=slug)
+        if request.method == 'GET':
+            post.unlike.add(request.user)
+            post.like.remove(request.user)
+            post.save()
+            return Response({'response': 'you unlike this post'}, status=status.HTTP_200_OK)
